feat: split hero content into a different component
